+++ src/sdk/modules/validators/smartSessions/decorators/mee/grantMeePermission.ts
@@ -1,5 +1,5 @@
 import type { Address, Prettify, PublicClient } from "viem"
-import { erc20Abi, parseUnits } from "viem"
+import { erc20Abi, getAbiItem, parseUnits, toFunctionSelector } from "viem"
 import type { BaseMeeClient } from "../../../../../clients/createMeeClient"
 import type { FeeTokenInfo } from "../../../../../clients/decorators/mee"
 import {
@@ -105,11 +105,31 @@ export const grantMeePermission = async <
     maxPaymentAmount = parseUnits("5", decimals)
   }
 
-  const grantPermissionParameters = actions.map((action) => {
-    const chainId = action.chainId
-    const actionTarget = action.actionTarget
+  /**
+   * Actions can be defined for multiple chains.
+   * Curernt approach is to build a single session for each unique chainId.
+   * So all the actions for a given chainId will be added to the same session.
+   *
+   * TODO: In future, we may want to add an additional parameter to the
+   * action object, which will define, whether the action can be batched into the same session
+   * or should be added to a separate session.
+   * This will also require changing the algorithm when using the permissions,
+   * since it will involve having several sessions on the same chain =>
+   * so it will require a proper algorithm of choosing which session we are using
+   * w/o requesting dev to provide the sessionId explicitly.
+   */
+
+  const uniqueChainIds = Array.from(
+    new Set(actions.map((action) => action.chainId))
+  )
+
+  const grantPermissionParameters = uniqueChainIds.map((chainId) => {
     const deployment = account.deployments.find(
       (deployment) => deployment?.client?.chain?.id === chainId
+    )
+
+    const actionsForChain = actions.filter(
+      (action) => action.chainId === chainId
     )
 
     const defaultVersionConfig: MEEVersionConfig =
@@ -118,24 +138,44 @@ export const grantMeePermission = async <
       deployment?.version.validatorAddress ||
       defaultVersionConfig.validatorAddress
 
-    const paymentActionPolicy =
-      feeToken && feeToken.chainId === chainId
-        ? {
-            actionTarget: feeToken.address,
-            actionTargetSelector: "0xa9059cbb" as Address, // transfer
-            actionPolicies: [
-              getPolicyForPayment(maxPaymentAmount!, feeToken.address)
-            ]
-          }
-        : undefined
+    let paymentAction: (ActionData & { chainId: number }) | undefined =
+      undefined
+    // if the fee token is involved in the permissions, try adding the payment action policy
+    if (feeToken && feeToken.chainId === chainId) {
+      // if some permission is already defining the policy for the feeToken.transfer, throw an error
+      if (
+        actionsForChain.some(
+          (action) =>
+            action.actionTargetSelector ===
+              toFunctionSelector(
+                getAbiItem({ abi: erc20Abi, name: "transfer" })
+              ) && action.actionTarget === feeToken.address
+        )
+      ) {
+        throw new Error(`You are defining the policy that prevents using ${feeToken.address} on chain ${chainId} as the fee token. 
+                         Possible solutions:
+                         1. Remove the 'transfer' method of ${feeToken.address} from the actions array.
+                         2. Use a different fee token.
+                         3. Use sponsored mode.`)
+      }
+
+      // else add the payment action policy
+      paymentAction = {
+        actionTarget: feeToken.address,
+        actionTargetSelector: "0xa9059cbb" as Address, // transfer
+        actionPolicies: [
+          getPolicyForPayment(maxPaymentAmount!, feeToken.address)
+        ],
+        chainId
+      }
+    }
 
     return {
       account: deployment,
       redeemer,
-      actions: [
-        ...actions.map((action) => ({ ...action, actionTarget })),
-        ...(paymentActionPolicy ? [paymentActionPolicy] : [])
-      ],
+      actions: paymentAction
+        ? [...actionsForChain, paymentAction]
+        : actionsForChain,
       sessionValidator: meeValidatorAddress,
       sessionValidatorInitData: redeemer, // initdata for the k1Mee validator is just the signer address
       permitERC4337Paymaster: true