+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # @biconomy/abstractjs
 
+## 1.1.9
+
+### Patch Changes
+
+- Patch: Smart sessions enhancement to have session batching. All the actions will be grouped into one session on the same chain.
+
+## 1.1.8
+
+### Patch Changes
+
+- Patches included:
+
+  1. Fixed Smart session mode usage
+  2. Added a universal policy to support payment token with warning
+
 ## 1.1.7
 
 ### Patch Changes